+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Current
 - Added backend level configuration ``FS_{BACKEND_NAME}_{KEY}``
 - Improved backend documentation
 - Use setuptools entry points to register backends.
+- Added `NONE` extensions specification
 
 0.3.0 (2017-03-05)
 ------------------